Score system nchallah cv

+++ src/main/java/Controllers/SeasonController.java
@@ -39,7 +39,7 @@ public class SeasonController {
     }
 
 
-    public static long StreamAverageScore(Season season) throws SQLException, IOException {return SeasonService.StreamAverageScore(season);}
+    public static Double StreamAverageScore(Season season) throws SQLException, IOException {return SeasonService.StreamAverageScore(season);}
 
     public static long StreamSumViewNumber(Season season) throws SQLException, IOException {return SeasonService.StreamSumViewNumber(season);}
